requirements.txt
selenium==3.141.0
gevent==21.12.0
flask
requests
multiprocess
schedule
sqlalchemy==1.4.7
pymysql
pandas==1.1.5
kafka-python==2.0.2